backend/requirements.txt
fastapi==0.102.0
uvicorn==0.23.2
pydantic==2.7.1
pydantic-settings==2.11.0
python-jose==3.3.0
httpx==0.25.0
python-dotenv==1.0.0
google-auth==2.21.0
requests==2.32.0
sqlalchemy==2.0.20
asyncpg
pyjwt==2.9.0
google-cloud-storage==2.10.0
python-multipart